% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WMisc.R
\docType{package}
\name{Wmisc}
\alias{Wmisc}
\alias{Wmisc-package}
\title{Wmisc -- Misc. helper functions and tools.}
\description{
This package contains a set of helper functions and tools that I couldn't find elsewhere or where I needed some slightly different behaviour.
}
\section{File access}{


Currently only one tool for file access is implemented: The \link{Tokenizer} to read string tokens from a (text) file.
}

